requirements.txt
# HTTP / async
aiohttp==3.9.5
aiohttp-retry==2.8.3

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1

# Data / analysis
numpy==1.26.4
scipy==1.13.0
pandas==2.2.2
pydantic==2.7.1
pydantic-settings==2.2.1

# Monitoring
prometheus-client==0.20.0

# Terminal UI
rich==13.7.1

# Utilities
python-dotenv==1.0.1
tenacity==8.3.0
structlog==24.1.0
httpx==0.27.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
respx==0.21.1

# Dev
ruff==0.4.4
mypy==1.10.0
